Move projectors into separate package
